The z-sorter is getting somewhere now!  The simplex seems to work for feasible start problems (I'll have to allocate a bit of extra memory to make infeasible start possible).
However, it is assumed that no facets intersect.  First of all, facet consist of triangles, and it would be nice to work on triangle level instead of facet level, because then it will be easier to think of how to deal with intersections and cyclic locking.

It should be added that the Makefile for the simlpex tester is not movable, and will only work on taran.  There is a nice example for the simplex tester, though, check out testers/opt2 (to be added in a minute!).




git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/drool/trunk@325 c1f300ed-6a01-0410-bd5d-8e39f192a3d9