I think the z sorter is working up to cyclic overlaps now, but it is slow!
The best demo of the z sorter at this time is the problems/kjmrobot.drool example.

I've also added a piece of new syntax.  It is an evaluated cut with just one binding, and fairly self-explanatory, see the curry.drool example, where a line almost like this one can be found:
  stderr << [foo [...] b:2 [...] 1 [...] 3]
It means
  stderr << [[[[foo b:2 ...] 1 ...] 3 ...]]



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/drool/trunk@331 c1f300ed-6a01-0410-bd5d-8e39f192a3d9