States seem useful now.  I will postpone the design and implementation of dynamic states until I get a better feeling for what one actually needs and what makes sense.
I have changed the name of the objects that spawn fresh states.  They are now called something beginning with "new" instead of "Hot".  "Hot" was impossible now when that would indicate a type name (it might turn out that a type can be meaningful here, but that is to be explored when typing is the topic), but I don't really like "new" either.  The problem with "new" is that it is a verb, while the value it binds to should not be thought of as a function; it is a very special object which the system knows how to use in order to get a fresh state.  It can't be a function, since functions can only return values.

Before updating all the examples to use states, I will take a look at the multiple return value stuff.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/drool/trunk@433 c1f300ed-6a01-0410-bd5d-8e39f192a3d9