It is possible that the document outline is useful now!
I hade to change the order tags are searched in GroupPair2D and GroupPair3D in order for the outline items to come in the logical order.

I admit that testing is totally incomplete.  I have only tested very few of the possible combinaitons of arguments to <destinations>.  Named destinations have not been tested at all, but if link annotations are added, they are likely to be so in the future.

It seems like Preview.app does not recognize the open/closed status of outline items, but it works in Adobe Reader.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/drool/trunk@631 c1f300ed-6a01-0410-bd5d-8e39f192a3d9