Introduce configure/automake for program compilation
This introduces the general and standardized configure/automake tool
chain to build Shapes.  So far only shapes and the unbalanced programs
are built.  Work remains to clean up the generated source and check
for dependencies of "external" includes.  Furthermore, this will not
produce an accpetable tar-ball for shipping since many headers would
be missing.

No attempt so far to generate documentation, nor to handle tests.


git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@798 c1f300ed-6a01-0410-bd5d-8e39f192a3d9