Introduce test for FLEX v. 2.5.33
The code uses new features in FLEX ruling out at least v 2.5.4.  At
the moment v. 2.5.33 is demanded, however it would probably be
possible to lower this somewhat if it turns out to be a problem for
people.

Since the distribution ships with generated scanner code a normal user
need not have FLEX installedm unless he/she plans to change any .ll
files.  Therefore, the tests does not halt nor even give an warning
when a proper FLEX isn't installed.  If the user doesn't read the
configure output carefully, the first indication of failur is when
FLEX is needed.  There are commented code to give a loud warning
during the configure phase, that could easily be activated if that
behavior is asked for.


git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@809 c1f300ed-6a01-0410-bd5d-8e39f192a3d9