Introduce test for BISON v. 2.3
The project uses new constructs in BISON that rules out old versions.
(No test has been performed to see exacly which version is needed so
the requirement v. 2.3 could very probably be too hard.)

Since the distribution ships with generated parser code a normal user
need not have BISON installedm unless he/she plans to change any .yy
files.  Therefore, the tests does not halt nor even give an warning
when a proper BISON isn't installed.  If the user doesn't read the
configure output carefully, the first indication of failur is when
BISON is needed.  There are commented code to give a loud warning
during the configure phase, that could easily be activated if that
behavior is asked for.


git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@810 c1f300ed-6a01-0410-bd5d-8e39f192a3d9