Proposing new command line file specifications.
This change originated from the need not to create files in the same directory as the input file.  Now, created files go into a directory specified using --odir, and defaults to "./".  This also implies a new interpretation of the last command line argument; it is no longer primarily a specification of the name base, but of the input file, and then the name base is derived form the input file (ignoring directory part of the file name path).

It is also possible to place the files produced for label generation in another directory, using the --tmpdir option or the SHAPESTMPDIR environment variable.  This feature is not yet tested.  Still, the label database (saving labels from one run to the next) is placed in the output directory, so it would be OK to set --tmpdir to a directory where files may be lost.

To only place the produced graphics in a particular directory, and all other files somewhere else, one may use the --out option together with the --odir option.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@822 c1f300ed-6a01-0410-bd5d-8e39f192a3d9