New names of many simple constants in the global environment.
Some time ago, values such as RGB_RED were renamed due to the limitation to lower case initials of variables.  Now when this restriction is removed, names are changed back.

It is expected that some examples do not fully reflect all changes in the global environment, but this is to be corrected when problems are encountered.  Of course, all tests have been updated as required and pass.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@845 c1f300ed-6a01-0410-bd5d-8e39f192a3d9