Update tests machinery
Complete work through of how tests are implemented.  The test
scripts now generate a log file for each test, and if called
with an argument > 0 will output the log on the terminal.

When testing graphical output, the pdf is no longer first converted
to a png and then compared.  Instead compare (part of the same
program suite as convert) is used to compare pdf files.  This
gives some more freedom since it does not demand a specific
graphics backend.


git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@861 c1f300ed-6a01-0410-bd5d-8e39f192a3d9