Rule to put the correct version string in configure.ac.
To synchronize configure.ac with VERSION, do
  make version_sync
Note, though, that this will only produce the right thing immediately after a version has been tagged.  This is explained in VERSION.

To conclude, do
  make version_sync
immediately after tagging, but before doing
  make distcheck
in order to ensure that the distributed program has the right version number in all places.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@869 c1f300ed-6a01-0410-bd5d-8e39f192a3d9