Relative LaTeX includes.
When using the --tmpdir option to place files related to the production of labels in a different directory than that where the Shapes source is locaed, care must be taken to ensure that .sty-files located relative to the Shapes source are still found.  At the moment, this is done using the TEXINPUTS environment variable, used by kpsewhich.  Although the tests go through now, the solution is not perfect for several reasons, including:
o  TEXINPUTS is not used by all LaTeX implementations.
o  At the moment, the only recognized relative path is ".".



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@874 c1f300ed-6a01-0410-bd5d-8e39f192a3d9