Make adding colors use fuzzy comparisions
This fixes a floating point problem that sometime can make correct colors
appear to be invalid, eg,

  (1-c)*RGB_RED + c*RGB_YELLOW

which by construct should be ok for c in [0, 1], turned out to render an
overflow in the red component.  The fuzzy tolerance is now set to 0.001
but it should be possible to lower this if floating point errors is the
only problem, now we help the sloppy user a bit too by being forgiving.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>


git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@947 c1f300ed-6a01-0410-bd5d-8e39f192a3d9