Optimization: use empty( ) instead of size( ) to test for emptyness
empty( ) is a O(1) operation in STL container, whereas size( ) may be
O(N) even though it is expected to be O(1) in most cases.  Still,
using empty( ) is more clear on the purpose of the test and sometimes
a bit faster.

The patch was prepered running:
	git grep -I -e "size( ) == 0" -e "size( ) != 0" -e "size( ) > 0"
and then going through the obtained list.


git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@1110 c1f300ed-6a01-0410-bd5d-8e39f192a3d9