Finally solved the design issue with procedure application.
The problem was to determine where procedures may be applied, so that the functional abstraction cannot be broken by procedure applications.  Basically, it is OK to call a procedure if it would be legal to inline its body at the place of the call.  The problem is solved by a runtime check of the relation between the procedure's environment and the environment in scope at the place of the call.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@1197 c1f300ed-6a01-0410-bd5d-8e39f192a3d9