Two things:  1) Using -output-directory option to pdfLaTeX.  2) New command line syntax for boolean options.
1)
By using -output-directory, we can now change directory to the source directory before running pdfLaTeX.  This way, we don't have to do funny things with $TEXINPUTS in order to find style files located relative to the Shapes source, and the simplicity of the current solution should make it more transparent for the user.

2)
Boolean-valued options are now specified like, for instance, --backtrace=yes or --backtrace=off.  Valid strings are
  yes, true, on, no, false, off
At the moment, "=yes" is not implied if one only writes --backtrace, but this is likely to be the case in the future.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@1212 c1f300ed-6a01-0410-bd5d-8e39f192a3d9