The directory for temporary files can be created automatically, if allowed.
To turn on the permission to created the directory for temporary files, one has to use the intentionally tricky command line option
  --tmp*=yes
(note that the asterix will have to be escaped in typical shells).

The file permissions of the created directories are taken from the deepest existing directory.  Shapes does not allow creation of root level directories, since that seems very stupid to do for temporary files.



git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/shapes/trunk@1232 c1f300ed-6a01-0410-bd5d-8e39f192a3d9