Reimplemented versioning
After changing to git the SVN based version information seized to
work.  This reimplements the same functionality with git tools
instead.  The reference to SVN revisions is droped, and a system to
indicate distance from the latest release is introduced.

The generated version number now have the following syntax:

    V[-N-gSHA1][-dirty]

V       is the closest annotated tag (set for each released version,
        eg, 0.3.1)
N       Number of commits since V
SHA1 4  first digits in the commit SHA1
dirty   indicate that there are uncommitted changes

The given date is either the date of the present commit, or current
time if it is a dirty state.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>