Fix race condition in version generation
A race condition appeared in the SHAPES-VERSION-GEN when the
repository was in a dirty state.  Each time SHAPES-VERSION-GEN was run
it updated the compilation time, and when the makefile was slower than
a second between runs this became an endless loop.  The solution taken
here is to simply ignore the second part of the timestamp.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>