Have configure find SAXON
configure is now used to determine where SAXON is to be found.  In the
following order, this is how SAXON is found (only the first hit is
tested):

  --with-saxon-path=path       given to configure
  $SAXON_PATH                  in the environment

As a last resort try without any path to SAXON.

If SAXON is not found, print a message at the end of the configure output
that it won't be possible to make the documentation.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>