Make Automake aware of doc/parts
This is the first step towards using Automake to build the docs in
doc/parts.  So far it is only aware of the directory, and uses
handcrafted Makefiles.

To do this the nasty self inclusion is broken apart, and noop check and
install targets are introduced not to halt the make process.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>