New feature --split=(no/flat/dir) allows document to be split into several output files.
This feature required rather big changes in the design of PDF generation.  The result may not be optimal, but reflects the PDF logic better now than before.  By not being optimal, I refer in particular to the empty slots in the cross reference table.

The meaning of the <output filename> varies depending on the split mode.  If --split=no there's no change, but if --split=flat it is just a filename prefix, and if --split=dir it is the name of a directory.  The default value is changed accordingly.

The document outline is ignored when the document is split, and invalid cross references are simply ignored.  However, page labels are respected.  In the future, splitting based on page labels should be considered.

We no longer delete output files in case of error.  However, when splitting the document into several files, files with confusing filenames will be deleted before creating any new files.

One of the most striking changes is that of global variables.  The old Kernel::the_pdfo served too many purposes and was removed, and replaced by three objects:
o Kernel::theIndirectObjectCount -- Used to assign a unique number to each generated indirect object.
o Kernel::theDocInfo -- Information about the document, to appear in all output files.
o Kernel::thePDFImporter -- Takes care of pages included from other PDF files.

Unfortunately, there's a fairly big bunch of new lines in the global function main.  Most of these deal with creation of output files, but there are also changes implied by the conceptual change from always-single-file to single-or-several-files.

New indirect objects are now created using the global function SimplePDF::indirect, and this caused minor changes in many files.

Each generated file is created by top-down analysis of the objects in the document catalog, hence ensuring that only objects that are used (or seem to be in use judging from XObject dictionaries and such) are written.  This may not lead to a great reduction in file size, but indicates better code.