Added the point-painting function "spot".
The function is implemented as stroking a singleton, closed, path with a round cap style.  This is very cheap compared to filling a circular path (which would use nine coordinate pairs instead of just one), but unfortunately not all PDF viewers do the obvious thing with singleton paths; the following list shows the compatibility with a few popular viewers:

  Preview.app 3.0.9:  yes
  xpdf 3.02:          no
  Adobe Reader 8.1.1: yes
  ImageMagic 6.2.9:   yes  (used to convert PDF to JPG for the documentation)

For compatibility, the command line option --spot-pair=yes may be given.  This makes spot generate a path between two point at the same location instead of the closed singleton path, and at least this works fine with xpdf 3.02.

I am not sure what the PDF standard says about singleton paths.  Anyone?