Compacting cref tables in the output.
Now, the xref tables contain no free slots at all.

This makes most sense when a big document catalog is split into separate output files for each page, as only a small fraction of the total number of indirect objects will be used in each output file.

The time and space penalties due to the extra layer of indirection should not be noticeable.