No longer confusing directory names with input filenames.
This became important now that --split=dir may create a directory named with just the filename base.  Hence, if there is an input file named <animation.shape>, shell completion will take you here if the output directory already exists:
  shapes --split=dir animation
and we want this to be a valid reference to the input file.

The risk that this will cause confusion in other situations is considered small.