Check for NaN signal at configure
Shapes doesn't work without a NaN signal, hence it is kind of rude to
not tell the user about that before she spent time to make and install
the program.  With this patch a test is introduced to the configure
script, and if the feature is missing the configure script halts.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>