Introduce SSI dependency generation
The program can now be called with the --deps command line option, to
produce a list of depenencies.  The SSI dependencies are the files that
are needed during the server-side include process, that is, files that
are included in the main file using the #include comment understood by
this program.