SSI processor: Let leading '/' indicate absolute search path.
This introduces the syntax of a leading '/' to indicate that a file is
absolute, that is, it will not be prepended with the directory of the
including file.

As a result of this patch SSIScanner::includeFilename_ becomes a
std::string as that is easier to handle, and less errorprone for
memory leaks.