Improved naming of log file when the base name is not defined.
Now, we try to deduce the name of the log file from the name of the output file.  If that fails, we use the base name, if defined.  As a last resort, we use the name "#shapes.log", in the output directory.

If one does not want the log file to appear in the output directory, one can always put is somewhere else using the --debug-log <filename> command line option.