No longer using html/head/base/@href in HTML.
This feature of HTML only supports absolute paths according to the
standard.  It was used with relative paths here, which worked fine
with Safari, but turned out not to work with FireFox.  Of course, we
want to follow standards, so the solution had to be changed.

The new solution is much more fragile since there are many places
where the base/@href path shall be added, and it must never be added
twice.

This solution is implemented in the XSL files.  Another, possibly more
robust, solution would have been to add yet another processing step,
similar to how we scan the HTML files for files they reference in
order to find build targets.