Resolved some "fixme" comments.
At some point during the development of the documentation build process, a few references had to be commented out to make things compile without adding files which were not directly related to the development of the build process itself.  Now that the build process is working and has been merged with the parallel development of the documentation content, these references could be safely re-introduced.