Abandoned use of absolute paths during documentation build.
This acutally consists of two small steps.

First, the current working directory is no longer prepended to the path to the main file passed to the SSI processor.  This behavior can be changed by passing "--absolute=yes" to the SSI processor.

Second, include filenames beginning with "^/" are interpreted as relative to the current working directory rather than relative to the including file.  The prefix "^/" is removed instead of prepending the string with the directory of the including file.

The choice of the prefix "^/" was made not to interfere with common shells' handling of special characters.  In particular, earlier ideas such as "<>/", were rejected for this reason.