Help to find installed documentation.
The installed documentation can now be pointed out by the compiler, by passing the --which-doc command line option.

The location of the installed documentation is also mentioned in the "See also" section of the man page.

To make it possible to request filenames without having to specify a base name or an output name, a dummy output name is now assumed when necessary.  (This only happens when there are filename requests, which ensures that the dummy name will never actually be used, since the program terminates after having responded to all filename requests.)