New test for graphics allows for several correct results.
If the same source produces slightly different results on different platforms (which is most likely to depend on differencies in TeX), it is possible to add all valid results to the set of acceptable results.  Any file with name matching {FNAME}.correct*.pdf is considered a correct result.

The test used in case the compare program is not available was also changed.  Now, we only check that the source compiles without errors in this case.  If an error occurs, this is indicated by touching out/{FNAME}.error, which is then detected by the test script generated from testTEMPLATE.