Testing the version number of ImageMagick.
This is to ensure that the compare program can be called with the -metric option.

The ./configure option --without-compare may be passed in order to prevent tests from being evaluated using compare.  This may be desirable if the compare tests turn out too picky, seeing differences in the graphics that the human eye doesn't care about, and which are not really signs of program error.

I am not sure which is the first version where -metric was available, but at least it is there in 6.4.0.