Try to make Makefiles more standardized
Make use of Automake hooks and locals in order to build things, this makes
the make files more standardized, and hence hopefully easier to maintain.

Also try to make use of as many standard constructions as possible,
eg, for installing new directories and data files.

Make sure not to forget to let users move their installation using the
$(DESTDIR) variable.  NB.  A remaining problem is how to make sure
the right resources are accessed by the installed shapes binary.