Doc: New design, refactoring, and lots of documentation for blockdraw_wfo.
Very nice table for dynamic variables.

The use of a table saves lots of space, and looks vastly nicer than the non-table solution used elsewhere.

The good thing is that the table uses the same kind of dynamic-variable elements as is used in other places, so the only difference is that the variable should be wrapped inside a <dynamic-variable-table>.

Refactoring CSS.

I also did some refactoring to avoid code duplication in the extensions/formats/ , since the formats for elements in extensions should look the same as the formats for elements in the core.