Fix the uninstall target
As part of the distcheck target, a test is performed that the program is
correctly uninstalled.  Make sure this test is passed.

At the moment, some files are left in the build tree that should probably
not be there, however, for the time being we choose to ignore them.
Consider trying to get rid of the $distuninstallcheck_listfiles definition.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>