Introducing exit codes.
The source should not contain any calls to <exit> with a number literal from now on.  Parts of the code which should not use the Shapes-specific exitcodes.h, should throw a const char * or a std::string instead of calling <exit>, so that the caller have the chance to use of of Shapes' exit codes.

The set of exit codes should be considered preliminary for some time... and perhaps it would be a good idea to build exitcodes.h from doc/parts/man/exit-codes.sxml to avoid out-of-synch situations between the two.