Doc: Minimize use of the customized dist and install design
By introducing another dependency file it is possible to get rid of
most of the customized dist and install elements.  Doing this makes
the Makefile a lot easier to maintain, at a small cost of a new file.
The procedure could possibly be streamlined even more.  This change at
the same time fixes the problem with files being left after uninstall.

Signed-off-by: Gustaf Hendeby <hendeby@isy.liu.se>