New feature: bleed boxes, and @TeX_bleed to control them.
Bleed boxes in Shapes are like bounding boxes, but are currently only used by the kernel to determine the visible area on the page.

The feature was implemented as a response to TeX sometimes producing bounding boxes that were slightly too small to fit the glyph.  This caused glyphs to get cropped if they appeared at the border of an image.  Now, @TeX_bleed may be used to make the bleed box a little bit larger than the bounding box, hence preventing ugly cropping.

It's funny, though, that with my present install of TeX, I can't find glyphs that need this feature.  I'm positive that, for instance a math mode $0$ is cropped when using some TeX distributions.

Anyway, the default value for @TeX_bleed is just 0.5bp, so I don't think this will be a problem even if all the glyph bounding boxes are acutally correct.

The functions bbox and bboxed were changed to take an additional (optional) argument, to control which box type to operate on.