Import of SVG paths.
The function <svg_path> takes the <d> string of an SVG <path> element, and interprets it using scaling only.  A more complete approach would also compute a shift, but this would be much more cumbersome to use for simple applications.  Please refer to the documentation of <svg_path> for some more details on this.

<svg_path> may not be extremely robust to errors in input data, but some measures are taken.

The result may either be a Path or a MultiPath value, depending on the number of sub-paths in the stirng.  Singleton paths may be disregarded.