Introducing fields of transforms.
The fields are not strictly necessary, since the parts they access can be computed using just transform application and arithmetics.  For instance, the linear part of T can be computed as (here, in 2D)
  L: T * [shift ~[T (0m,0m)]]
Then, given the linear part L, the "first column", T.Lx, may be computed as
  [L (1m,0m)] / 1m
and so forth.  However, the above expressions obscure simple abstractions such as "the linear part of T", which is why the new fields are introduced.  Now, everyone can write T.L instead, which is both easier to write, and easier to recognize.