Mutators for Catalog states.
While working on this, there was a problem with order of initialization of globals.  Problem was solved by making the initialization order explicit, by listing objects for initialization in main.cc.  This requires an ugly const_cast.

However, if one remembers to test new methods being added, one will immediately detect if methods have not been registered.  Hence, there is no real danger in requiring all warm types with methods to be listed in main.cc.