Template base class for methods of core types.
The new class template should be used as base class for all methods defined for types in the core language.  This will both reduce the amount of code that has to be written for each new method, and make standardize how methods present themseleves in error messages by initializing the title_ variable according to the examples given by current methods.