Standardizing methods of core types.
This should not change the language in any aspect, but deals only with how methods are implemented.

The idea is that methods are a type property, which allows them to be "stored" in the type.  Well, actually, the methods themselves are not stored in the type, since they depend on the object they belong to, what is stored in the type is the method factory.

The code has been rewritten so that the name of a method appears only once, and so that the member data <title_> may be used in the implementation of a method to present itself correctly in error messages.

Existing types with methods have been rewritten using the new technique, and shall serve as examples when methods are added to other types as well.