Removing "named let", and traces of other "let" constructs.
Since the named let was not used anywhere in the examples, I decided to remove it.  If one want's it back later, one may always take the code for it prior to this commit.

Anyway, the Shapes syntax for a named let was a rather odd bird, since there was no ordinary let to generalize.  Note also, that the effect of a named let can easily be achieved in Shapes, for instance:
factorial: \ n0 .>
  { a : \ n:n0 -> [if n=='0 '1 n*[a n-'1]]
	  [a] }