Bugfix: Source locations over ##needs seem OK now.
The implemented fix was to let ##needs and ##include return a T_srcLoc first of all.  This is a valid GroupElem in the parser (hence, ##needs and ##include can only appear as group elements now, which may be considered a new bug), which ensures that the location of the Group non-terminal gets correct.  When the Group is turned into an Ast::CodeBracket, the dummy nodes are removed.