Including FlexLexer.h in distribution.
This is not only necessary if the user does not have Flex installed, but more confusingly, if the user has the wrong version of Flex installed.

Note that while FlexLexer.h is included in the distribution, it is not included in the repository.  Developers need to have a proper version of Flex installed, and should rely on the FlexLexer.h that comes with it.