Bugfix: Solving trouble with proj-web linking to HTML documentation.
The basic problem is that the proj-web index.html links, with @href='doc/', to the HTML documentation.  This is not meant to be regarded as a dependency while building, as proj-web and the HTML documentation should be pretty much independent.  However, the xhtml-deps filter was picking up this directory.

A symlink from doc/build/proj-web to doc/build/html was introduced for convenience, so that the link with @href="doc/" can be resolved while developing proj-web.

I tried to include the symlink in the nobas_dist_html_DATA variable in proj-web/Makefile.am, but this was not treated as just a symlink during install.  Another install hook would have been needed to treat it correctly, and I didn't want to (didn't know how to) do this.

Instead, I chose to change the xhtml-deps filter, to exclude any reference that ends with '/'.  Since one generally doesn't know the name of the index file in a directory, the correct treatment of such references would be just to ensure that the directory exists, but let's hope that the directories we need contain files that will be installed, hence causing the directories themselves to be created even if we don't explicitly demand this.

For some reason, we get some kind of cyclic dependency chain (which make doesn't detect) if $(stylesdir) is a relative reference.  A comment was left to prevent ths error to be re-introduced.