Bugfix: Exponents in scientific notation.
Exponents in scientific notation were lost, so the bug caused a number such as "1.5*^3" to evaluate to 1.5.  Lengths were also affected.

Note that the scanner must be prepared that the name of a length may conflict with the "e" used by strtod and similar functions.  Hence, the digits in the exponent must be detected manually.  For instance, note that
  1.5e4
should scan as the length 1.5e, followed by the real number 4.