Bugfix: Confusing to pass dynamic argument twice to the tack-on methods of states.
The dynamic environment used to be passed both as a separate argument, but also as part of evalState.  While using the separately passed dynamic environment used to be the correct choice, the mere possibility of using evalState->dyn_ is a serious problem.  Indeed, most tack-on operations incorrectly used evalState->dyn_, which of course showed as a bug.

Now, the separately passed dynamic environment has been removed, and all tack-on operations use evalState->dyn_ instead.