Revised operator precedence rules.
This change introduces severe backward incompatibilities.  Sources cannot be updated using simple regexp matching.

The motivations behind the new precedence rules include:
o  Dynamic bindings too often required parentheses when put in scope.  Fix: give | higher precedence than -> .
o  Functions using real currying, for instance
     \ a -> \ b -> 10 * a + b
   need a left associative [] operator.  Fix: change from right to left associative.
o  With [] (and [!]) now being left associative, they can be given the same precedence level as [...] and [!...], reducing the number of precedence levels for the poor user to keep track of.
o  What about convenient expressions such as
     stroke [] [shift p] [] immerse [] [circle 1cm]
   ?  Fix: introduce new operator which acs like the // operator i Mathematica.  The new operator has the notation >> , and is left associative.
   The convenient expression changes to
     [circle 1cm] >> immerse >> [shift p] >> stroke
   which has the advantage over the original expression that reading the expression from left to right is closer to how we think about it;
     "First create a circle of radius 1cm, then immerse it, then shift it to p, and finally stroke it."
o  The xor operator is not associative.  Hence it cannot have the same precedence as the or operator.  Fix: use same precedence rule as is found in MySQL, that is, <xor> is in between <or> and <and>.
o  Keep the number of precedence levels given to =, /=, <, >, <=, >= to a minimum.  Fix: put <= and >= on the same level as = and /=.  Note that < and > cannot have the same level, as these need to be left associative (generally, the comparison operators should not be associative at all), and must also have lower precedence than --.  Since we want the precedence of -- to be the same as that of ++ due to symmetry, and close to other arithmetic operators, having difference precedence rules for < and > than the other comparison operators also allows the number of strange precedence levels to be kept at a minimum.

Recall that Shapes is a special-purpose language for graphics -- this is why a small number of strange precedence rules is tolerated for the purpose of getting a nice syntax for path construction.  If a dedicated path construction operator is introduced in the future, the precedence level of ++ and -- may be changed to something more natural, and the precedence level of < and > may be changed to something next to that of the other comparison operators.