Bugfix: Filters on user dynamic variables, and taking away old warning.
The problem with the filters was that they shoudl be forced.  The fix causes also the default value to be forced, but this can easily be changed in the future.

The warning warned about something which seems like the correct thing to do, so I removed it.  However, the warning has probably never been seen by anyone, since the present bugfix was required to produce it at this time.

A test was added to ensure that one does not accidentaly change back to passing the filter as a thunk.