Examples: Updated to compile with new associativity of [].
The new operator >> was used instead of [] where found appropriate.  Many uses of [] were left without modification.  In particular, [] is still used when just applying a transform to an object.

When several transforms are combined using the * operator, I find it a bit confusing to use the >> operator, so I tend to either use the basic function application syntax or the [] operator in such cases.