Doc: Reminding that escaping ` is not required to produce the left double quote in LaTeX.
Rather than using poor man's strings or data strings, one may enter the left double quote directly in a normal string.

This is illustrated with a new example in the tutorial.

The example is rather important, since it discusses an exception to the design goal that LaTeX source should need no special treatment inside Shapes string litterals.

Note that the acute accent by itself is currently not supported by the utf8.def inputenc definition in LaTeX.  Hence, the following will not work:
  (TeX \gThis is a string in Shapes\gblah\G\G)
where I used \g for the grave accent, and \G for the right accent, just to make this commit message readble as ASCII.  Since LaTeX changes the meaning of the grave accent to a LEFT SINGLE QUOTATION MARK, this example would miss the point even if the input was accepted.